Move hook into hooks folder
